🎨 Style: Layout 구성
