+++ src/app/(header)/(narrow-view)/cart/_components/cartFooter/cartFooterType.ts
