+++ src/app/(header)/(narrow-view)/cart/_components/cartGroup/cartGroup.stories.tsx
@@ -1,15 +1,14 @@
 import React, { useState } from 'react';
-import type { StoryObj } from '@storybook/react';
-import { TcartInfo } from './_components/cartItem/cartItemType';
-import { ThandleSelectItem } from './cartType';
-import CartTitle from './_components/cartTitle';
-import CartGroup from './_components/cartGroup';
-import CartItem from './_components/cartItem';
-import CartFooter from './_components/cartFooter';
-import CartLayout from './layout';
+import type { Meta, StoryObj } from '@storybook/react';
+import CartGroup from '.';
+import { ThandleSelectItem } from '../../cartType';
+import { TcartInfo } from '../cartItem/cartItemType';
+import CartItem from '../cartItem';
+import Layout from '../../../layout';
 
 const meta = {
-  title: 'cart/Cart',
+  title: 'cart/CartGroup',
+  component: CartGroup,
   tags: ['autodocs'],
   parameters: {
     layout: 'fullscreen',
@@ -17,24 +16,44 @@ const meta = {
       appDirectory: true,
     },
     componentSubtitle:
-      'Cart page에 사용되는 컴포넌트를 조합하는 컨테이너입니다 ',
+      'CartGroup는 CartItem을 그룹으로 묶어 제어하는 컴포넌트입니다. ',
+    docs: {
+      description: {
+        component: `
+- CartGroup의 Checkbox를 통해 선택과 해제가 가능합니다.\n
+- CartGroup의 삭제 아이콘을 통해 장바구니에서 상품 삭제가 가능합니다.\n
+`,
+      },
+    },
   },
-};
+  argTypes: {
+    children: {
+      description: 'CartItem 컴포넌트를 의미합니다',
+    },
+    selectAll: {
+      description: 'CartGroup의 Checkbox의 onChange 속성에 할당 될 함수입니다.',
+    },
+    isAllSelected: {
+      description:
+        'CartGroup의 Checkbox의 isChecked 속성에 할당 될 함수로 boolean 값을 반환합니다',
+    },
+  },
+} satisfies Meta<typeof CartGroup>;
 
 export default meta;
 
-export const ExampleCart: StoryObj = {
+export const ExampleCartGroup: StoryObj<typeof CartGroup> = {
   decorators: [
     (Story) => (
-      <CartLayout>
+      <Layout>
         <Story />
-      </CartLayout>
+      </Layout>
     ),
   ],
   parameters: {
     docs: {
       description: {
-        story: '기본으로 사용되는 Cart입니다.',
+        story: '기본으로 사용되는 default CartGroup입니다.',
       },
     },
   },
@@ -93,30 +112,14 @@ export const ExampleCart: StoryObj = {
       return selectedItems.filter((item) => item.cart_id === id).length !== 0;
     };
 
-    const calculateTotalPrice = () => {
-      return selectedItems.reduce(
-        (totalPrice, items) => totalPrice + items.cart_price,
-        0,
-      );
-    };
-
     return (
-      <>
-        <CartTitle />
-        <CartGroup selectAll={selectAll} isAllSelected={isAllSelected}>
-          {data.map((item) => (
-            <CartItem
-              data={item}
-              handleSelectItem={handleSelectItem}
-              isSelected={isSelected}
-            />
-          ))}
-        </CartGroup>
-        <CartFooter
-          totalPrice={calculateTotalPrice()}
-          selectedItemsLength={selectedItems.length}
+      <CartGroup selectAll={selectAll} isAllSelected={isAllSelected}>
+        <CartItem
+          data={data[0]}
+          handleSelectItem={handleSelectItem}
+          isSelected={isSelected}
         />
-      </>
+      </CartGroup>
     );
   },
 };
