+++ src/app/(header)/(narrow-view)/cart/_components/cartItem/cartItemType.ts
