+++ src/app/(header)/(narrow-view)/cart/_components/cartTitle/cartTitle.module.scss
@@ -1,8 +1,3 @@
-.frame {
-  max-width: 48.125rem;
-  margin: 1.25rem auto 0 auto;
-}
-
 .title {
   width: 100%;
   display: flex;
