+++ src/app/(header)/(narrow-view)/layout.module.scss
@@ -1,0 +1,4 @@
+.frame {
+  max-width: 770px;
+  margin: 0 auto;
+}
