+++ src/app/(header)/(narrow-view)/layout.tsx
@@ -1,7 +1,8 @@
 import React from 'react';
+import styles from './layout.module.scss';
 
 function Layout({ children }: { children: React.ReactNode }) {
-  return <>{children}</>;
+  return <main className={styles.frame}>{children}</main>;
 }
 
 export default Layout;
