+++ src/app/(header)/(wide-view)/layout.module.scss
@@ -1,0 +1,4 @@
+.frame {
+  max-width: 1080px;
+  margin: 0 auto;
+}
