+++ src/app/(header)/(wide-view)/layout.tsx
@@ -1,0 +1,8 @@
+import React from 'react';
+import styles from './layout.module.scss';
+
+function Layout({ children }: { children: React.ReactNode }) {
+  return <main className={styles.frame}>{children}</main>;
+}
+
+export default Layout;
