+++ src/app/(header)/(wide-view)/rooms/[id]/page.tsx
