--- src/app/(header)/cart/layout.tsx
@@ -1,8 +1,0 @@
-import React from 'react';
-import styles from './cart.module.scss';
-
-function CartLayout({ children }: { children: React.ReactNode }) {
-  return <main className={styles.frame}>{children}</main>;
-}
-
-export default CartLayout;
